refactor: migration static-resource to core-browser
