+++ packages/express-file-server/package.json
@@ -17,9 +17,9 @@
     "url": "git@github.com:opensumi/core.git"
   },
   "dependencies": {
+    "@opensumi/ide-core-browser": "workspace:*",
     "@opensumi/ide-core-common": "workspace:*",
     "@opensumi/ide-core-node": "workspace:*",
-    "@opensumi/ide-static-resource": "workspace:*",
     "koa-mount": "^4.0.0"
   },
   "devDependencies": {
