+++ packages/express-file-server/src/browser/file-server.contribution.ts
@@ -3,7 +3,7 @@ import { Domain, URI, AppConfig, Schemes } from '@opensumi/ide-core-browser';
 import {
   StaticResourceContribution,
   StaticResourceService,
-} from '@opensumi/ide-static-resource/lib/browser/static.definition';
+} from '@opensumi/ide-core-browser/lib/static-resource/static.definition';
 
 import { EXPRESS_SERVER_PATH } from '../common';
