+++ packages/extension-manager/package.json
@@ -15,7 +15,6 @@
     "@opensumi/ide-core-common": "workspace:*",
     "@opensumi/ide-core-node": "workspace:*",
     "@opensumi/ide-extension": "workspace:*",
-    "@opensumi/ide-static-resource": "workspace:*",
     "compressing": "^1.6.2",
     "fs-extra": "^10.1.0",
     "node-fetch": "^2.6.7",
