+++ packages/extension/__tests__/browser/main.thread.extensions.test.ts
@@ -2,6 +2,7 @@ import { Injectable, Injector } from '@opensumi/di';
 import { RPCProtocol } from '@opensumi/ide-connection/lib/common/rpcProtocol';
 import { IContextKeyService, AppConfig } from '@opensumi/ide-core-browser';
 import { MockedStorageProvider } from '@opensumi/ide-core-browser/__mocks__/storage';
+import { StaticResourceService } from '@opensumi/ide-core-browser/lib/static-resource';
 import * as ideCoreCommon from '@opensumi/ide-core-common';
 import { IReporter, DefaultReporter } from '@opensumi/ide-core-common';
 import { AppConfig as NodeAppConfig } from '@opensumi/ide-core-node';
@@ -10,7 +11,6 @@ import { IExtensionStorageService } from '@opensumi/ide-extension-storage';
 import { FileSearchServicePath } from '@opensumi/ide-file-search';
 import { MockFileServiceClient } from '@opensumi/ide-file-service/__mocks__/file-service-client';
 import { OutputPreferences } from '@opensumi/ide-output/lib/browser/output-preference';
-import { StaticResourceService } from '@opensumi/ide-static-resource/lib/browser';
 import { IGlobalStorageServer } from '@opensumi/ide-storage';
 import { IThemeService, IIconService } from '@opensumi/ide-theme';
 import { IconService } from '@opensumi/ide-theme/lib/browser';
