+++ packages/extension/__tests__/browser/main.thread.workspace.test.ts
@@ -15,6 +15,7 @@ import {
 } from '@opensumi/ide-core-browser';
 import { injectMockPreferences } from '@opensumi/ide-core-browser/__mocks__/preference';
 import { useMockStorage } from '@opensumi/ide-core-browser/__mocks__/storage';
+import { StaticResourceService } from '@opensumi/ide-core-browser/lib/static-resource';
 import {
   Uri as vscodeUri,
   Emitter,
@@ -82,7 +83,6 @@ import { FileService, FileSystemNodeOptions } from '@opensumi/ide-file-service/l
 import { DiskFileSystemProvider } from '@opensumi/ide-file-service/lib/node/disk-file-system.provider';
 import { MonacoService } from '@opensumi/ide-monaco';
 import MonacoServiceImpl from '@opensumi/ide-monaco/lib/browser/monaco.service';
-import { StaticResourceService } from '@opensumi/ide-static-resource/lib/browser';
 import { StaticResourceServiceImpl } from '@opensumi/ide-static-resource/lib/browser/static.service';
 import { IWebviewService } from '@opensumi/ide-webview';
 import { IWorkspaceService } from '@opensumi/ide-workspace';
