+++ packages/extension/package.json
@@ -23,12 +23,12 @@
   },
   "dependencies": {
     "@opensumi/ide-connection": "workspace:*",
+    "@opensumi/ide-core-browser": "workspace:*",
     "@opensumi/ide-core-node": "workspace:*",
     "@opensumi/ide-debug": "workspace:*",
     "@opensumi/ide-file-search": "workspace:*",
     "@opensumi/ide-file-service": "workspace:*",
     "@opensumi/ide-logs": "workspace:*",
-    "@opensumi/ide-static-resource": "workspace:*",
     "@opensumi/ide-task": "workspace:*",
     "@opensumi/ide-terminal-next": "workspace:*",
     "@opensumi/ide-webview": "workspace:*",
