+++ packages/quick-open/package.json
@@ -24,7 +24,6 @@
     "@opensumi/ide-core-browser": "workspace:*",
     "@opensumi/ide-dev-tool": "workspace:*",
     "@opensumi/ide-monaco": "workspace:*",
-    "@opensumi/ide-static-resource": "workspace:*",
     "@opensumi/ide-theme": "workspace:*",
     "@opensumi/ide-utils": "workspace:*",
     "@opensumi/ide-workspace": "workspace:*"
