+++ packages/quick-open/src/browser/quick-open-item.service.ts
@@ -1,7 +1,7 @@
 import { Injectable, Autowired } from '@opensumi/di';
 import { IQuickPickItemButtonEvent, QuickInputButton } from '@opensumi/ide-core-browser/lib/quick-open';
+import { StaticResourceService } from '@opensumi/ide-core-browser/lib/static-resource';
 import { Emitter, Event } from '@opensumi/ide-core-common';
-import { StaticResourceService } from '@opensumi/ide-static-resource/lib/browser';
 import { IconType, IIconService, IThemeService } from '@opensumi/ide-theme';
 
 import { iconPath2URI } from '../common/icon';
