+++ packages/startup/src/browser/common-modules.ts
@@ -28,7 +28,6 @@ import { PreferencesModule } from '@opensumi/ide-preferences/lib/browser';
 import { QuickOpenModule } from '@opensumi/ide-quick-open/lib/browser';
 import { SCMModule } from '@opensumi/ide-scm/lib/browser';
 import { SearchModule } from '@opensumi/ide-search/lib/browser';
-import { StaticResourceModule } from '@opensumi/ide-static-resource/lib/browser';
 import { StatusBarModule } from '@opensumi/ide-status-bar/lib/browser';
 import { StorageModule } from '@opensumi/ide-storage/lib/browser';
 import { TaskModule } from '@opensumi/ide-task/lib/browser';
@@ -53,7 +52,6 @@ export const CommonBrowserModules: ConstructorOf<BrowserModule>[] = [
   ExplorerModule,
   FileTreeNextModule,
   FileServiceClientModule,
-  StaticResourceModule,
   SearchModule,
   FileSchemeModule,
   OutputModule,
