+++ packages/theme/package.json
@@ -23,7 +23,6 @@
     "@opensumi/ide-core-browser": "workspace:*",
     "@opensumi/ide-dev-tool": "workspace:*",
     "@opensumi/ide-file-service": "workspace:*",
-    "@opensumi/ide-logs": "workspace:*",
-    "@opensumi/ide-static-resource": "workspace:*"
+    "@opensumi/ide-logs": "workspace:*"
   }
 }
