+++ packages/theme/src/browser/icon-theme-data.ts
@@ -1,7 +1,7 @@
 import { Injectable, Autowired } from '@opensumi/di';
 import { getDebugLogger, URI, parseWithComments, formatLocalize } from '@opensumi/ide-core-browser';
+import { StaticResourceService } from '@opensumi/ide-core-browser/lib/static-resource';
 import { IFileServiceClient } from '@opensumi/ide-file-service';
-import { StaticResourceService } from '@opensumi/ide-static-resource/lib/browser';
 
 import { IIconTheme } from '../common';
