+++ packages/theme/src/browser/icon.service.ts
@@ -14,7 +14,7 @@ import {
   ExtensionDidContributes,
   GeneralSettingsId,
 } from '@opensumi/ide-core-browser';
-import { StaticResourceService } from '@opensumi/ide-static-resource/lib/browser';
+import { StaticResourceService } from '@opensumi/ide-core-browser/lib/static-resource';
 
 import {
   ThemeType,
