+++ packages/webview/package.json
@@ -25,7 +25,6 @@
     "@opensumi/ide-core-electron-main": "workspace:*",
     "@opensumi/ide-dev-tool": "workspace:*",
     "@opensumi/ide-editor": "workspace:*",
-    "@opensumi/ide-static-resource": "workspace:*",
     "@opensumi/ide-theme": "workspace:*"
   },
   "peerDependencies": {
